requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dotenv==0.0.5
Flask==1.1.2
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Fredyma-Studio-Project==1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
twilio==6.38.1
urllib3==1.25.9
Werkzeug==1.0.1
